feat: refactor code to reuse existing sdk
