+++ lib/bento/sdk/version.rb
@@ -1,5 +1,5 @@
 module Bento
   class Analytics
-    VERSION = "0.1"
+    VERSION = "0.2.0"
   end
 end
