--- lib/bento_v2/client.rb
@@ -1,7 +1,0 @@
-module Bento
-  class Client
-    def initialize(config = Configuration.new)
-      @config = config
-    end
-  end
-end
