--- lib/bento_v2/version.rb
@@ -1,3 +1,0 @@
-module Bento
-  VERSION = '0.0.1'.freeze
-end
